+++ packages/config-loader/src/loader.ts
@@ -15,20 +15,18 @@
  */
 
 import fs from 'fs-extra';
-import { resolve as resolvePath, dirname } from 'path';
+import { resolve as resolvePath, dirname, isAbsolute } from 'path';
 import { AppConfig, JsonObject } from '@backstage/config';
-import {
-  resolveStaticConfig,
-  readConfigFile,
-  readEnvConfig,
-  readSecret,
-} from './lib';
+import { readConfigFile, readEnvConfig, readSecret } from './lib';
 
 export type LoadConfigOptions = {
-  // Root paths to search for config files. Config from earlier paths has lower priority.
-  rootPaths: string[];
+  // The root directory of the config loading context. Used to find default configs.
+  configRoot: string;
 
-  // The environment that we're loading config for, e.g. 'development', 'production'.
+  // Absolute paths to load config files from. Configs from earlier paths have lower priority.
+  configPaths: string[];
+
+  // TODO(Rugvip): This will be removed in the future, but for now we use it to warn about possible mistakes.
   env: string;
 
   // Whether to read secrets or omit them, defaults to false.
@@ -77,13 +75,37 @@ export async function loadConfig(
   options: LoadConfigOptions,
 ): Promise<AppConfig[]> {
   const configs = [];
+  const { configRoot } = options;
+  const configPaths = options.configPaths.slice();
 
-  const configPaths = await resolveStaticConfig(options);
+  // If no paths are provided, we default to reading
+  // `app-config.yaml` and, if it exists, `app-config.local.yaml`
+  if (configPaths.length === 0) {
+    configPaths.push(resolvePath(configRoot, 'app-config.yaml'));
+
+    const localConfig = resolvePath(configRoot, 'app-config.local.yaml');
+    if (await fs.pathExists(localConfig)) {
+      configPaths.push(localConfig);
+    }
+
+    const envFile = `app-config.${options.env}.yaml`;
+    if (await fs.pathExists(resolvePath(configRoot, envFile))) {
+      console.error(
+        `Env config file '${envFile}' is not loaded as APP_ENV and NODE_ENV-based config loading has been removed`,
+      );
+      console.error(
+        `To load the config file, use --config <path>, listing every config file that you want to load`,
+      );
+    }
+  }
 
   try {
     const secretPaths = new Set<string>();
 
     for (const configPath of configPaths) {
+      if (!isAbsolute(configPath)) {
+        throw new Error(`Config load path is not absolute: '${configPath}'`);
+      }
       const config = await readConfigFile(
         configPath,
         new Context({