+++ plugins/catalog-backend/src/ingestion/processors/BitbucketApiReaderProcessor.ts
@@ -15,9 +15,9 @@
  */
 
 import { LocationSpec } from '@backstage/catalog-model';
-import { Config } from '@backstage/config';
-import fetch, { HeadersInit, RequestInit } from 'node-fetch';
+import fetch from 'cross-fetch';
 import * as result from './results';
+import { Config } from '@backstage/config';
 import { CatalogProcessor, CatalogProcessorEmit } from './types';
 
 // ***********************************************************************
@@ -70,8 +70,8 @@ export class BitbucketApiReaderProcessor implements CatalogProcessor {
       const response = await fetch(url.toString(), this.getRequestOptions());
 
       if (response.ok) {
-        const data = await response.buffer();
-        emit(result.data(location, data));
+        const data = await response.text();
+        emit(result.data(location, Buffer.from(data)));
       } else {
         const message = `${location.target} could not be read as ${url}, ${response.status} ${response.statusText}`;
         if (response.status === 404) {